+++ src/snowflake/snowpark/_internal/analyzer/select_statement.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
+# Copyright (c) 2012-2024 Snowflake Computing Inc. All rights reserved.
 #
 
 from abc import ABC, abstractmethod