+++ CHANGELOG.md
@@ -26,8 +26,8 @@
     - current_session
     - date_trunc
     - udf
-    - current_database
-    - current_session
+    - object_construct
+    - object_construct_keep_null
 - Added the function `DataFrame.write.csv` to unload data from a ``DataFrame`` into one or more CSV files in a stage.
 - Added distributed tracing using open telemetry apis for action functions in `DataFrame` and `DataFrameWriter`:
   - snowflake.snowpark.DataFrame:
@@ -38,10 +38,12 @@
     - show
   - snowflake.snowpark.DataFrameWriter:
     - save_as_table
+- Added support for snow:// URLs to `snowflake.snowpark.Session.file.get` and `snowflake.snowpark.Session.file.get_stream`
 
 ### Bug Fixes
 
 - Fixed a bug in local testing that null filled columns for constant functions.
+- Fixed a bug causing `snowflake.snowpark.Session.file.get_stream` to fail for quoted stage locations
 
 ## 1.14.0 (2024-03-20)